:bug: fix: remove highlight wrapper
