+++ src/Markdown/wrapper.tsx
@@ -1,15 +1,17 @@
 import CopyButton from '@/components/CopyButton';
 import { DownOutlined, RightOutlined } from '@ant-design/icons';
-import { ActionIcon, Button, Select, type SelectProps } from '@ant-design/pro-editor';
+import { ActionIcon, Button, Highlight, Select, type SelectProps } from '@ant-design/pro-editor';
 import classNames from 'classnames';
 import { memo, useState } from 'react';
-import { DivProps, Flexbox } from 'react-layout-kit';
-import { HighlightBase, HighlightProps } from './defalut';
-import { languageMap } from './hooks/useHighlight';
+import { Flexbox } from 'react-layout-kit';
 import { useStyles } from './style';
-import { THEME_LIGHT } from './theme';
 
-export interface HighlighterWrapperProps extends DivProps {
+const options: SelectProps['options'] = Object.keys(languageMap).map((key) => ({
+  label: key,
+  value: key.toLowerCase(),
+}));
+
+export interface HighlightWrapperProps {
   /**
    * @description The code content to be highlighted
    */
@@ -18,34 +20,26 @@ export interface HighlighterWrapperProps extends DivProps {
    * @description The language of the code content
    */
   language?: string;
+  /**
+   * classNames
+   */
+  className?: string;
+  /**
+   * style
+   */
+  style?: React.CSSProperties;
 }
-const options: SelectProps['options'] = Object.keys(languageMap).map((key) => ({
-  label: key,
-  value: key.toLowerCase(),
-}));
 
-export const FullFeatureWrapper = memo<HighlighterWrapperProps & HighlightProps>((props) => {
-  const {
-    children,
-    language = 'markdown',
-    className,
-    style,
-    theme = THEME_LIGHT,
-    copyable = true,
-    type = 'block',
-  } = props || {};
+const HighlightWrapper = memo((props: HighlightWrapperProps) => {
+  const { children, language = 'markdown', className, style } = props || {};
   const [expand, setExpand] = useState(true);
   const [lang, setLang] = useState(language);
-  const { styles } = useStyles({
-    theme,
-    type,
-  });
+  const { styles } = useStyles();
 
   return (
     <div className={classNames(styles.wrapper, className)} style={style}>
       <Flexbox align={'center'} className={styles.header} horizontal justify={'space-between'}>
         <ActionIcon
-          className={styles.expland}
           icon={expand ? <DownOutlined size={14} /> : <RightOutlined size={14} />}
           onClick={() => setExpand(!expand)}
           size={24}
@@ -70,17 +64,15 @@ export const FullFeatureWrapper = memo<HighlighterWrapperProps & HighlightProps>
           suffixIcon={false}
           value={[lang.toLowerCase()]}
         />
-        {copyable && <CopyButton className={styles.copy} content={children} />}
+        <CopyButton className={styles.copy} content={children} />
       </Flexbox>
-      <HighlightBase
-        {...props}
-        language={lang?.toLowerCase()}
-        style={expand ? {} : { height: 0, overflow: 'hidden', border: 'none' }}
-        copyable={false}
-        showLanguage={false}
-      />
+      {expand ? (
+        <Highlight language={lang?.toLowerCase()} copyable={false} showLanguage={false}>
+          {children}
+        </Highlight>
+      ) : null}
     </div>
   );
 });
 
-export default FullFeatureWrapper;
+export default HighlightWrapper;
